+++ dtu_hpc_cli/run.py
@@ -4,8 +4,11 @@ from dtu_hpc_cli.client import get_client
 from dtu_hpc_cli.config import cli_config
 
 
-def execute_run(commands: list[str]):
+def execute_run(arguments: list[str]):
+    if len(arguments) == 0:
+        rich.print("[bold red]No command provided.")
+        return
+
+    command = " ".join(arguments)
     with get_client() as client:
-        for command in commands:
-            rich.print(f"[bold blue]> {command}")
-            client.run(command, cwd=cli_config.remote_path)
+        client.run(command, cwd=cli_config.remote_path)